requirements.txt
asgiref==3.9.1
category_encoders==2.8.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
joblib==1.5.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
patsy==1.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.5
threadpoolctl==3.6.0
tzdata==2025.2
